% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia1_series.R
\name{eia1_series}
\alias{eia1_series}
\alias{eia1_series_req}
\title{Request Data Using Legacy APIv1 Series ID Compatibility Endpoint}
\usage{
eia1_series(series_id)

eia1_series_req(series_id)
}
\arguments{
\item{series_id}{EIA APIv1 series ID (length 1 character)}
}
\value{
tibble with series data, or for eia1_series_req a request object
}
\description{
Request Data Using Legacy APIv1 Series ID Compatibility Endpoint
}
\examples{
\dontrun{
eia1_series("ELEC.SALES.CO-RES.A")
}

}
