% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia2-package.R
\docType{package}
\name{eia2-package}
\alias{eia2-package}
\title{eia2: Wrapper for the US Energy Information Administration API Version 2}
\description{
Provides API access to the open data API from the US Energy Information Administration (EIA) \url{https://www.eia.gov/opendata/}. Use of the API requires a free API key. The package includes functions to request data using API parameters including data columns, facets, start and end dates, and sorting. The version 2 API provides a compatibility endpoint for retrieving data using version 1 API series ID's.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jameelalsalam.github.io/eia2/}
}

}
\author{
\strong{Maintainer}: Jameel Alsalam \email{jalsalam@gmail.com} (\href{https://orcid.org/0000-0001-9801-5558}{ORCID})

Other contributors:
\itemize{
  \item Matthew Leonawicz \email{mfleonawicz@gmail.com} (\href{https://orcid.org/0000-0001-9452-2771}{ORCID}) [contributor]
}

}
\keyword{internal}
