% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{eia2}
\alias{eia2}
\title{Retrieve data from the EIA API}
\usage{
eia2(
  route = "",
  facets = list(),
  data_cols = character(),
  frequency = NULL,
  start = NULL,
  end = NULL,
  sort = data.frame(column = "period", direction = "desc"),
  offset = 0,
  length = 5000,
  out = "json",
  api_key = eia_get_key()
)
}
\arguments{
\item{route}{to base dataset}

\item{facets}{list of facet filter specifications}

\item{data_cols}{names of data columns to retrieve}

\item{frequency}{data frequency}

\item{start}{filter start of data requested}

\item{end}{filter end of data requested}

\item{sort}{data frame specifying sorting}

\item{offset}{starting position, for paged results}

\item{length}{number of data points to retrieve}

\item{out}{output data format, "json" or "xml"}

\item{api_key}{character api key}
}
\description{
Retrieve data from the EIA API
}
\examples{
\dontrun{
eia2_req("electricity/retail-sales", data_cols = "price")
}

}
