% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-big.R
\name{eia2_data_big}
\alias{eia2_data_big}
\title{Retrieve data from the EIA API in Multiple Requests}
\usage{
eia2_data_big(
  route = "",
  facets = list(),
  data_cols = character(),
  frequency = NULL,
  start = NULL,
  end = NULL,
  sort = data.frame(column = "period", direction = "desc"),
  offset = 0,
  length = 5000,
  by = 5000,
  out = "json",
  api_key = eia_get_key()
)
}
\arguments{
\item{route}{to base dataset}

\item{facets}{list of facet filter specifications}

\item{data_cols}{names of data columns to retrieve}

\item{frequency}{data frequency}

\item{start}{filter start of data requested}

\item{end}{filter end of data requested}

\item{sort}{data frame specifying sorting}

\item{offset}{starting position, for paged results}

\item{length}{number of data points to retrieve}

\item{by}{number of data points in each page}

\item{out}{output data format, "json" or "xml"}

\item{api_key}{character api key}
}
\description{
Retrieve data from the EIA API in Multiple Requests
}
\examples{
\dontrun{
eia2("electricity/retail-sales", data_cols = "price", length = 10)
}

}
\seealso{
\link{request}, \code{\link[=eia2_req_perform]{eia2_req_perform()}}, and \code{\link[=eia2_resp_data]{eia2_resp_data()}} provide
a stepwise workflow to build a request, perform it, and extract formatted data.
}
