% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets.R
\name{eia2_facet}
\alias{eia2_facet}
\title{Determine Valid Values of a Facet}
\usage{
eia2_facet(dataset, facet)
}
\arguments{
\item{dataset}{route to EIA dataset}

\item{facet}{id of facet to query for valid values (length 1 character)}
}
\value{
tibble of valid facet values (id & name)
}
\description{
Determine Valid Values of a Facet
}
