% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_perform.R
\name{eia2_req_perform}
\alias{eia2_req_perform}
\title{Perform an eia2 request}
\usage{
eia2_req_perform(
  req,
  path = NULL,
  verbosity = NULL,
  mock = getOption("httr2_mock", NULL),
  rate = getOption("eia2_rate", 4),
  api_key = eia_get_key()
)
}
\arguments{
\item{req}{An eia2 \link{request}.}

\item{path}{Optionally, path to save body of the response.}

\item{verbosity}{How much information to print? This is an integer from 0 to 3,
passed to \code{httr2::request_perform()}.}

\item{mock}{A mocking function. Passed to \code{httr2::request_perform()}.}

\item{rate}{Maximum request rate in requests per second, passed to \code{httr2::req_throttle()}.}

\item{api_key}{Your API authentication key.}
}
\description{
After preparing a \link{request}, call \code{eia2_req_perform()} to perform it, fetching
the results back to R as a \link{response}.

\code{eia2_req_perform()} wraps \code{httr2::req_perform()}. The api_key is not added
to the request object until the request is performed. Additionally, it
applies rate limit behavior to comply with API guidelines.
}
