% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{eia2_resp_data}
\alias{eia2_resp_data}
\title{Extract formatted data from responses}
\usage{
eia2_resp_data(eia_resp)
}
\description{
If you use the stepwise workflow of \code{eia2_request()} -> \code{eia2_req_perform()},
then the next step is to extract useful data from the responses.

For data requests, \code{eia2_resp_data()} returns body data as a tibble.

For metadata requests:
\itemize{
\item \code{eia2_resp_meta_routes()} returns metadata on child routes as a tibble.
\item \code{eia2_resp_meta_summary()} returns a summary metadata lists.
}
}
