% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_expand_params.R
\name{query_expand_params}
\alias{query_expand_params}
\title{Expand/Flatted Nested Objects to be Used as url query parameters}
\usage{
query_expand_params(params, auto_unbox = FALSE)
}
\arguments{
\item{params}{named list or character/numeric vector}

\item{auto_unbox}{whether to unbox lists}
}
\value{
a flat list of name-value pairs, whose values are length-1 atomic
}
\description{
Expand/Flatted Nested Objects to be Used as url query parameters
}
