% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{eia2_req}
\alias{eia2_req}
\title{Request data from EIA API version 2}
\usage{
eia2_req(
  route = "",
  facets = list(),
  data_cols = character(),
  frequency = NULL,
  start = NULL,
  end = NULL,
  sort = data.frame(column = "period", direction = "desc"),
  offset = 0,
  length = 5000,
  out = "json",
  api_key = NULL
)
}
\arguments{
\item{route}{to base dataset}

\item{facets}{list of facet filter specifications}

\item{data_cols}{names of data columns to retrieve}

\item{frequency}{data frequency}

\item{start}{filter start of data requested}

\item{end}{filter end of data requested}

\item{sort}{data frame specifying sorting}

\item{offset}{starting position, for paged results}

\item{length}{number of data points to retrieve}

\item{out}{output data format, "json" or "xml"}

\item{api_key}{character api key, or NULL to omit}
}
\description{
Request data from EIA API version 2
}
\examples{
eia2_req("electricity")
eia2_req("electricity/retail-sales")
eia2_req("electricity/retail-sales/data")
eia2_req("electricity/retail-sales/data", data_cols = "price")
eia2_req("electricity/retail-sales", data_cols = "price") # same as one above

}
