% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia2-package.R
\docType{package}
\name{eia2-package}
\title{eia2: Wrapper for the US Energy Information Administration API Version 2}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jameel Alsalam \email{jalsalam@gmail.com} (\href{https://orcid.org/0000-0001-9801-5558}{ORCID})

Other contributors:
\itemize{
  \item Matthew Leonawicz \email{mfleonawicz@gmail.com} (\href{https://orcid.org/0000-0001-9452-2771}{ORCID}) [contributor]
}

}
\keyword{internal}
